libksi: reverted pull request #536 for libksi 3.4.0.1
As there is currently an issue open for the libkis versioning,
the patch cannot be merged yet as it break compatibility with the
current release library.

For more see: https://github.com/rsyslog/rsyslog/pull/536

+++ runtime/librsksi_read.c
@@ -88,7 +88,7 @@ static inline void
 outputKSIHash(FILE *fp, char *hdr, const KSI_DataHash *const __restrict__ hash, const uint8_t verbose)
 {
 	const unsigned char *digest;
-	size_t digest_len;
+	unsigned digest_len;
 	KSI_DataHash_extract(hash, NULL, &digest, &digest_len); // TODO: error check
 
 	fprintf(fp, "%s", hdr);
